push t3

+++ part3/app/api/v1/reviews.py
@@ -16,17 +16,11 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_requiered
     def post(self):
         """Register a new review"""
+        current_user = get_jwt_identity()
         review_data = api.payload
-        place = facade.get_place(review_data['place_id'])
-        if not place:
-            return {'error': 'Place not found'}, 400
-        user = facade.get_user(review_data['user_id'])
-        if not user:
-            return {'error': 'User not found'}, 400
-        if place.owner.id == user.id:
-            return {'error': 'User cannot review their own place'}, 400
         try:
             new_review = facade.create_review(review_data)
             return new_review.to_dict(), 201
@@ -53,13 +47,17 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_requiered
     def put(self, review_id):
         """Update a review's information"""
+        current_user = get_jwt_identity()
         review_data = api.payload
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
         
+        if review.user.id != current_user:
+            return {'error':'Unauthorized action'}, 403
         try:
             facade.update_review(review_id, review_data)
             return {'message': 'Review updated successfully'}, 200
@@ -68,11 +66,16 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_requiered()
     def delete(self, review_id):
         """Delete a review"""
+        current user = get_jwt_identity()
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if review.user.id != current_user:
+            return {'error': 'Unauthorized action'}, 403
         
         try:
             facade.delete_review(review_id)
